+++ handler/confirmemail.go
@@ -6,7 +6,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//    http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,8 +16,6 @@
 package handler
 
 import (
-	"encoding/json"
-	"fmt"
 	"net/http"
 	"time"
 
@@ -35,6 +33,8 @@ type ConfEmailLink struct {
 }
 
 func Confirmemail(w http.ResponseWriter, r *http.Request) {
+
+	t := RequestInit(r)
 	//Log Request
 	//1. Collect need data
 	var userip = sf.ReadUserIP(r)
@@ -45,14 +45,7 @@ func Confirmemail(w http.ResponseWriter, r *http.Request) {
 
 	//1. We get request with email and hash and check if it need data exist
 	if r.URL.Query()["token"][0] == "" || r.URL.Query()["email"][0] == "" {
-
-		errormsg := []sf.ErrorMsg{}
-		errorans := sf.ErrorMsg{
-			Code:    "000001",
-			Message: "Missing Token or Email",
-		}
-		errormsg = append(errormsg, errorans)
-		sendanswer(w, r, ans, errormsg)
+		errorAnswer(w, r, t, 400, "000001", "Missing Token or Email")
 		return
 	}
 
@@ -79,33 +72,20 @@ func Confirmemail(w http.ResponseWriter, r *http.Request) {
 			sf.SetErrorLog("confirmemail.go: " + err.Error())
 		}
 		//return "error with db"
+		errorAnswer(w, r, t, 400, "000001", "DB Connection Error")
 	}
 
 	//4.1. Check if hash is exist in db users
 	rows := db.Conn.Where(`hash = ? and mail = ?`, data.Token, data.Email).First(&userHash)
 	if rows.RowsAffected == 0 {
 		// return error. Hash is not exist in db
-
-		errormsg := []sf.ErrorMsg{}
-		errorans := sf.ErrorMsg{
-			Code:    "000008",
-			Message: "Hash is not exist in db",
-		}
-		errormsg = append(errormsg, errorans)
-		sendanswer(w, r, ans, errormsg)
+		errorAnswer(w, r, t, 400, "000008", "Hash is not exist in db")
 		return
 	}
 
 	curtime := int(time.Now().Unix())
 	if userHash.Livetime < curtime {
-
-		errormsg := []sf.ErrorMsg{}
-		errorans := sf.ErrorMsg{
-			Code:    "000009",
-			Message: "Hash is overtime",
-		}
-		errormsg = append(errormsg, errorans)
-		sendanswer(w, r, ans, errormsg)
+		errorAnswer(w, r, t, 400, "000009", "Hash is overtime")
 		return
 	}
 
@@ -116,61 +96,7 @@ func Confirmemail(w http.ResponseWriter, r *http.Request) {
 	db.Conn.Delete(sf.TimeHash{}, "hash = ?", data.Token)
 
 	ans["response"] = "100002" // email confirmed
-	sendanswer(w, r, ans, nil)
+	moduleAnswerv3(w, r, ans, t)
 	return
-
-}
-
-func sendanswer(w http.ResponseWriter, r *http.Request, ans map[string]interface{}, errmsg []sf.ErrorMsg) {
-
-	if ans["error"] != nil {
-		var resp sf.ErrorResponse
-		resp.Success = 0
-		resp.Error = errmsg
-		var lg = fmt.Sprintf("%s", ans["lang"])
-		if lg == "" {
-			lg = "eng"
-		}
-		resp.Language = lg
-		resp.TimeStamp = int(time.Now().Unix())
-		answer, err := json.Marshal(resp)
-		if err != nil {
-
-			if viper.GetBool("server.sentry") {
-				sentry.CaptureException(err)
-			} else {
-				sf.SetErrorLog("confirmemail.go: " + err.Error())
-			}
-			return
-		}
-		w.Header().Set("Access-Control-Allow-Origin", "*")
-		w.Header().Set("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE")
-		w.Header().Set("Access-Control-Allow-Headers", "Authorization, Content-Type")
-		w.Header().Set("Server", "Gufo")
-		w.Header().Set("Content-Type", "application/json")
-		w.Write([]byte(answer))
-	} else {
-		var resp sf.SuccessResponse
-		resp.Success = 1
-		resp.Language = "eng"
-		resp.TimeStamp = int(time.Now().Unix())
-		resp.Data = ans
-		answer, err := json.Marshal(resp)
-		if err != nil {
-
-			if viper.GetBool("server.sentry") {
-				sentry.CaptureException(err)
-			} else {
-				sf.SetErrorLog("confirmemail.go: " + err.Error())
-			}
-			return
-		}
-		w.Header().Set("Access-Control-Allow-Origin", "*")
-		w.Header().Set("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE")
-		w.Header().Set("Access-Control-Allow-Headers", "Authorization, Content-Type")
-		w.Header().Set("Server", "Gufo")
-		w.Header().Set("Content-Type", "application/json")
-		w.Write([]byte(answer))
-	}
 
 }