+++ handler/processreq.go
@@ -31,69 +31,41 @@ import (
 
 	"github.com/getsentry/sentry-go"
 	sf "github.com/gogufo/gufo-api-gateway/gufodao"
-	"github.com/microcosm-cc/bluemonday"
+	pb "github.com/gogufo/gufo-api-gateway/proto/go"
 
 	"github.com/spf13/viper"
 )
 
-func ProcessREQ(w http.ResponseWriter, r *http.Request, version int) {
+func ProcessREQ(w http.ResponseWriter, r *http.Request, t *pb.Request, version int) {
 
-	t := &sf.Request{}
 	//Determinate plugin name, params etc.
 	//
 	path := r.URL.Path
 	patharray := strings.Split(path, "/")
 	pathlenth := len(patharray)
 
-	p := bluemonday.UGCPolicy()
+	//p := bluemonday.UGCPolicy()
 
 	if pathlenth < 3 {
 
-		if version == 3 {
-			nomoduleAnswerv3(w, r)
-		} else {
-			nomoduleAnswer(w, r)
-		}
+		errorAnswer(w, r, t, 401, "0000235", "Wrong Path Lenth")
 
 		return
 
 	}
 	//Plagin Name
-	t.Module = p.Sanitize(patharray[3])
-	t.Path = path
 
-	t.APIVersion = "v3"
+	vrs := "v3"
+	t.APIVersion = &vrs
 
 	if version == 2 {
-		t.APIVersion = "v2"
+		vrs = "v2"
+		t.APIVersion = &vrs
 	}
 
-	t.Method = r.Method
-
-	t.Sign = viper.GetString("server.sign")
-
-	t.IP = sf.ReadUserIP(r)
-
-	t.UserAgent = r.UserAgent()
-
-	if t.Module == "entrypoint" {
-		if version == 3 {
-			nomoduleAnswerv3(w, r)
-		} else {
-			nomoduleAnswer(w, r)
-		}
+	if *t.Module == "entrypoint" {
+		errorAnswer(w, r, t, 401, "0000235", "Wrong module")
 		return
-	}
-
-	//Function in Plugin
-	if pathlenth >= 5 {
-		t.Param = p.Sanitize(patharray[4])
-	}
-
-	//ID for function in plugin
-	if pathlenth >= 6 {
-		t.ParamID = p.Sanitize(patharray[5])
-
 	}
 
 	if r.Method == "POST" {
@@ -121,51 +93,34 @@ func ProcessREQ(w http.ResponseWriter, r *http.Request, version int) {
 				paramMap[k] = v[0]
 			}
 		}
-		t.Args = paramMap
+		anydt := sf.ToMapStringAny(paramMap)
+		t.Args = anydt
 
 	}
 
 	//check for session
 	t = checksession(t, r)
 
-	if t.UID != "" && t.Readonly == 1 {
-		if version == 3 {
-			nomoduleAnswerv3(w, r)
-		} else {
-			nomoduleAnswer(w, r)
-		}
+	if *t.UID != "" && *t.Readonly == int32(1) {
+
+		errorAnswer(w, r, t, 401, "0000235", "Read Only User")
 		return
+
 	}
 
-	mdir := viper.GetString("server.plugindir")
-	pluginname := fmt.Sprintf("microservices.%s", t.Module)
+	pluginname := fmt.Sprintf("microservices.%s", *t.Module)
 
 	if !viper.IsSet(pluginname) {
-		msg := fmt.Sprintf("No Module %s", t.Module)
+		msg := fmt.Sprintf("No Module %s", *t.Module)
 		if viper.GetBool("server.sentry") {
 			sentry.CaptureMessage(msg)
 		} else {
 			sf.SetErrorLog(msg)
 		}
-		if version == 3 {
-			nomoduleAnswerv3(w, r)
-		} else {
-			nomoduleAnswer(w, r)
-		}
+		errorAnswer(w, r, t, 401, "0000235", msg)
 		return
 	}
-
-	//Check is it plugin or GRPC server
-	plugintype := viper.GetString(fmt.Sprintf("%s.type", pluginname))
-
-	if plugintype == "plugin" {
-
-		file := viper.GetString(fmt.Sprintf("%s.file", pluginname))
-		mod := fmt.Sprintf("%s%s", mdir, file)
-		loadmodule(w, r, mod, t, version)
-	} else if plugintype == "server" {
-		//Load microservice
-		connectgrpc(w, r, t)
-	}
+	//Load microservice
+	connectgrpc(w, r, t)
 
 }